// web/src/api/generated/models/workflow-version-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 工作流版本历史响应
 */
export interface WorkflowVersionResponse {
    /**
     * 版本记录ID
     */
    'id'?: number;
    /**
     * 工作流ID
     */
    'workflowId'?: number;
    /**
     * 版本号
     */
    'version'?: number;
    /**
     * 名称快照
     */
    'name'?: string;
    /**
     * 描述快照
     */
    'description'?: string;
    /**
     * 工作流定义快照JSON
     */
    'definition'?: string;
    /**
     * 发布说明
     */
    'publishNote'?: string;
    /**
     * 发布者ID
     */
    'publishedBy'?: number;
    /**
     * 创建时间
     */
    'createTime'?: string;
}